# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API para consulta de codigos postales",
        "description": "Busca de codigos postales, información de estados, municipios y sus asentamientos",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://codigo-postal.test/api"
        }
    ],
    "paths": {
...
        "/asentamientos": {
            "get": {
                "tags": [
                    "asentamientos"
                ],
                "summary": "Listado de todos los asentamientos, con paginación 15 por página",
                "operationId": "129781c92fd8ceda6f25d8e31f373437",
                "parameters": [
                    {
                        "name": "filter[name]",
                        "in": "query",
                        "description": "Buscar por nombre del asentamiento",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenar por 'name', '-name'",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Cantidad de registros por consulta",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Si esta activada la paginación, indica a que página desea ir",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "cp": {
                                                        "type": "number",
                                                        "example": 97825
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Mukuiche"
                                                    },
                                                    "tipo_asentamiento": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "tipo": {
                                                                "type": "string",
                                                                "example": "Pueblo"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "ciudad": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "null"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "null"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "municipio": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Abalá"
                                                            },
                                                            "estado": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "number",
                                                                        "example": "1"
                                                                    },
                                                                    "nombre": {
                                                                        "type": "string",
                                                                        "example": "Yucatán"
                                                                    },
                                                                    "abrev": {
                                                                        "type": "string",
                                                                        "example": "Yuc."
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/asentamientos/{id}": {
            "get": {
                "tags": [
                    "asentamientos"
                ],
                "summary": "Obtener datos de un asentamiento por el id",
                "operationId": "06512656860114171eb043be5fb4bee3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Obtener un asentamiento por id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "cp": {
                                                        "type": "number",
                                                        "example": 97825
                                                    },
                                                    "nombre": {
                                                        "type": "string",
                                                        "example": "Mukuiche"
                                                    },
                                                    "tipo_asentamiento": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "tipo": {
                                                                "type": "string",
                                                                "example": "Pueblo"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "ciudad": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "null"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "null"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "municipio": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "number",
                                                                "example": "1"
                                                            },
                                                            "nombre": {
                                                                "type": "string",
                                                                "example": "Abalá"
                                                            },
                                                            "estado": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "number",
                                                                        "example": "1"
                                                                    },
                                                                    "nombre": {
                                                                        "type": "string",
                                                                        "example": "Yucatán"
                                                                    },
                                                                    "abrev": {
                                                                        "type": "string",
                                                                        "example": "Yuc."
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Recurso Suburb no encontrado, puede que el id no exista en la base de datos"
                                        },
                                        "code": {
                                            "type": "number",
                                            "example": "404"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}